Handle v2 temperature limits as half-degree values
Decode Koolnova v2 register 40076 temperature limit bytes as Celsius values stored in half-degree steps, and encode writes back to the raw Modbus representation.

Expose the related Home Assistant number entities with a 0.5°C step and adjusted maximum value, while rejecting values that are not aligned to half-degree increments.

Document the confirmed half-degree encoding for register 40076 in the README.

+++ custom_components/koolnova_bms/koolnova/device.py
@@ -852,8 +852,8 @@ class Koolnova:
         )
 
     async def async_set_v2_temperature_limits(self,
-                                                max_heating_limit:int,
-                                                min_cooling_limit:int,
+                                                max_heating_limit:float,
+                                                min_cooling_limit:float,
                                                 ) -> dict:
         ''' Set Koolnova v2 temperature limits '''
         return await self._async_write_v2_register(
